+++ example/publishing_resources/source/example.cpp
@@ -19,7 +19,7 @@ void post_method_handler( const shared_ptr< Session >& session )
 {
     const auto request = session->get_request( );
     
-    size_t content_length = 0;
+    int content_length = 0;
     request->get_header( "Content-Length", content_length );
     
     session->fetch( content_length, [ ]( const shared_ptr< Session >& session, const Bytes & body )
